# Conflicts:
#	DataAccessLayer/BusinessLogiLayer/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
#	DataAccessLayer/DataAccessLayer.v12.suo
#	DataAccessLayer/DataAccessLayer/DataAccessLayer.csproj
#	DataAccessLayer/DataAccessLayer/VBDQ.dbml
#	DataAccessLayer/DataAccessLayer/VBDQ.dbml.layout
#	DataAccessLayer/DataAccessLayer/VBDQ.designer.cs
#	DataAccessLayer/DataAccessLayer/obj/Debug/TempPE/VBDQ.designer.cs.dll
#	DataAccessLayer/PresentationLayer/MainForm.Designer.cs
#	DataAccessLayer/PresentationLayer/PresentationLayer.csproj

+++ DataAccessLayer/DTO/PhieuDichVu_DTO.cs
@@ -14,6 +14,6 @@ namespace DTO
         public string NgayGiao { get; set; }
         public string DiaChi { get; set; }
         public decimal TongTien { get; set; }
-        public decimal TinhTrang { get; set; }
+        public int TinhTrang { get; set; }
     }
 }
